requirements.txt
meshtastic>=2.3.0
ollama>=0.2.0
chromadb>=0.5.0
pyyaml>=6.0
# pymupdf>=1.24.0  # PDF support — deferred, optional
